move results_tmodel to results dir
