+++ statsmodels/miscmodels/tests/results/results_tmodel.py
