+++ statsmodels/miscmodels/tests/test_tmodel.py
@@ -145,7 +145,7 @@ class TestTModel(CheckTLinearModelMixin):
         modf = TLinearModel.from_formula("price ~ CRSP",
                                 data={"price":mm.m_marietta, "CRSP":mm.CRSP})
         resf = modf.fit(method='bfgs', disp=False)
-        from .results_tmodel import res_t_dfest as res2
+        from results.results_tmodel import res_t_dfest as res2
         cls.res2 = res2
         cls.res1 = res  # take from module scope temporarily
         cls.resf = resf
@@ -164,7 +164,7 @@ class TestTModelFixed(object):
                                 fix_df=3)
         resf = modf.fit(method='bfgs', disp=False)
         #TODO: no reference results yet
-        #from results_tmodel import res_t_dfest as res2
+        #from .results.results_tmodel import res_t_dfest as res2
         #cls.res2 = res2
         cls.res1 = res  # take from module scope temporarily
         cls.resf = resf
